+++ app/controllers/scientists_controller.rb
@@ -1,3 +1,5 @@
 class ScientistsController < ApplicationController
- 
+  def show 
+    @scientist = Scientist.find(params[:id])
+  end
 end
