+++ app/controllers/scientists_experiment.rb
@@ -1,0 +1,5 @@
+class ScientistsExperimentController < ApplicationController
+  def destroy 
+    
+  end
+end
